Revert "Merge branch 'legacy' of https://github.com/pjw6193/caliber into legacy"
This reverts commit 83a15f6febb170cef5d6433ae24a8b7c03ec8cce, reversing
changes made to abafee01ceef40648d1d23d1c5ed2c7ee160df0b.
